Changes to test-integration

+++ test-integration/scripts/10.1-install-jhipster-blueprint.sh
@@ -6,21 +6,27 @@ source $(dirname $0)/00-init-env.sh
 #-------------------------------------------------------------------------------
 # Install Blueprint
 #-------------------------------------------------------------------------------
-cd "$HOME"
 echo "*** generator-jhipster-blueprint: use local version at JHI_REPO=$JHI_REPO"
+if [[ "$INSTALL_BLUEPRINT" == "global" ]]; then
+    echo "Installing globally"
+    npm install -g "$JHI_HOME"
 
-cd "$JHI_HOME"
-git --no-pager log -n 10 --graph --pretty='%Cred%h%Creset -%C(yellow)%d%Creset %s %Cgreen(%cr) %C(bold blue)<%an>%Creset' --abbrev-commit
+else
+    cd "$HOME"
 
-rm -rf "$JHI_FOLDER_APP"
-mkdir -p "$JHI_FOLDER_APP"
-cd "$JHI_FOLDER_APP"
+    cd "$JHI_HOME"
+    git --no-pager log -n 10 --graph --pretty='%Cred%h%Creset -%C(yellow)%d%Creset %s %Cgreen(%cr) %C(bold blue)<%an>%Creset' --abbrev-commit
 
-# Create package.json otherwise npm will error and break
-# https://github.com/visionmedia/debug/issues/261
-npm init -f
+    mkdir -p "$JHI_FOLDER_APP"
+    cd "$JHI_FOLDER_APP"
 
-npm install "$JHI_HOME"
+    # Create package.json otherwise npm will error and break
+    # https://github.com/visionmedia/debug/issues/261
+    npm init -f
 
-# Remove package.json to let jhipster create it's own
-rm package.json
+    # Install blueprint
+    npm install "$JHI_HOME"
+
+    # Remove package.json to let jhipster create it's own
+    rm package.json
+fi
